refactoring marshal functions and files (#2)
* refactor: renaming marshaling/unmarshaling files and functions for both json and pongo schema

* renamed schema_marshal.go to schema_pongoschema_marshal.go
* removed schema_interface.go, split across schema.go, schema_jsonschema_marshal.go, schema_pongoschema_marshal.go
* renamed from DefaultSchemaUnmarshalMap to GlobalSchemaUnmarshalMap

* feat: added option to set the global mapper

* refactor: fixed renaming from UnmarshalSchemaWithMapper to UnmarshalPongoSchemaWithMapper

* testing: added tests for SetGlobalPongoSchemaUnmarshalMap

* refactor: renamed PongoSchemaUnmarshalMap to PongoSchemaUnmarshalMapper
